chatbot_pdf/requirements.txt
pymupdf
gradio
openai
python-dotenv
